% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MotionTemplate.R
\name{MotionTemplate}
\alias{MotionTemplate}
\title{<Add Title>}
\usage{
MotionTemplate(
  navigation = NULL,
  maincontet,
  width = "100\%",
  height = "100\%",
  elementId = NULL,
  ...
)
}
\description{
<Add Description>
}
